Move bind/unbind events functions to new file, with example usage.
