+++ core/bubble.js
@@ -12,6 +12,7 @@
 
 goog.provide('Blockly.Bubble');
 
+goog.require('Blockly.eventHandling');
 goog.require('Blockly.Scrollbar');
 goog.require('Blockly.Touch');
 goog.require('Blockly.utils');
@@ -66,14 +67,14 @@ Blockly.Bubble = function(
 
   /**
    * Mouse down on bubbleBack_ event data.
-   * @type {?Blockly.EventData}
+   * @type {?Blockly.eventHandling.Data}
    * @private
    */
   this.onMouseDownBubbleWrapper_ = null;
 
   /**
    * Mouse down on resizeGroup_ event data.
-   * @type {?Blockly.EventData}
+   * @type {?Blockly.eventHandling.Data}
    * @private
    */
   this.onMouseDownResizeWrapper_ = null;
@@ -137,14 +138,14 @@ Blockly.Bubble.ANCHOR_RADIUS = 8;
 
 /**
  * Mouse up event data.
- * @type {?Blockly.EventData}
+ * @type {?Blockly.eventHandling.Data}
  * @private
  */
 Blockly.Bubble.onMouseUpWrapper_ = null;
 
 /**
  * Mouse move event data.
- * @type {?Blockly.EventData}
+ * @type {?Blockly.eventHandling.Data}
  * @private
  */
 Blockly.Bubble.onMouseMoveWrapper_ = null;
@@ -155,11 +156,11 @@ Blockly.Bubble.onMouseMoveWrapper_ = null;
  */
 Blockly.Bubble.unbindDragEvents_ = function() {
   if (Blockly.Bubble.onMouseUpWrapper_) {
-    Blockly.unbindEvent_(Blockly.Bubble.onMouseUpWrapper_);
+    Blockly.eventHandling.unbind(Blockly.Bubble.onMouseUpWrapper_);
     Blockly.Bubble.onMouseUpWrapper_ = null;
   }
   if (Blockly.Bubble.onMouseMoveWrapper_) {
-    Blockly.unbindEvent_(Blockly.Bubble.onMouseMoveWrapper_);
+    Blockly.eventHandling.unbind(Blockly.Bubble.onMouseMoveWrapper_);
     Blockly.Bubble.onMouseMoveWrapper_ = null;
   }
 };
@@ -299,10 +300,10 @@ Blockly.Bubble.prototype.createDom_ = function(content, hasResize) {
   }
 
   if (!this.workspace_.options.readOnly) {
-    this.onMouseDownBubbleWrapper_ = Blockly.bindEventWithChecks_(
+    this.onMouseDownBubbleWrapper_ = Blockly.eventHandling.checkAndBind(
         this.bubbleBack_, 'mousedown', this, this.bubbleMouseDown_);
     if (this.resizeGroup_) {
-      this.onMouseDownResizeWrapper_ = Blockly.bindEventWithChecks_(
+      this.onMouseDownResizeWrapper_ = Blockly.eventHandling.checkAndBind(
           this.resizeGroup_, 'mousedown', this, this.resizeMouseDown_);
     }
   }
@@ -387,9 +388,9 @@ Blockly.Bubble.prototype.resizeMouseDown_ = function(e) {
       new Blockly.utils.Coordinate(
           this.workspace_.RTL ? -this.width_ : this.width_, this.height_));
 
-  Blockly.Bubble.onMouseUpWrapper_ = Blockly.bindEventWithChecks_(
+  Blockly.Bubble.onMouseUpWrapper_ = Blockly.eventHandling.checkAndBind(
       document, 'mouseup', this, Blockly.Bubble.bubbleMouseUp_);
-  Blockly.Bubble.onMouseMoveWrapper_ = Blockly.bindEventWithChecks_(
+  Blockly.Bubble.onMouseMoveWrapper_ = Blockly.eventHandling.checkAndBind(
       document, 'mousemove', this, this.resizeMouseMove_);
   Blockly.hideChaff();
   // This event has been handled.  No need to bubble up to the document.
@@ -821,10 +822,10 @@ Blockly.Bubble.prototype.setColour = function(hexColour) {
  */
 Blockly.Bubble.prototype.dispose = function() {
   if (this.onMouseDownBubbleWrapper_) {
-    Blockly.unbindEvent_(this.onMouseDownBubbleWrapper_);
+    Blockly.eventHandling.unbind(this.onMouseDownBubbleWrapper_);
   }
   if (this.onMouseDownResizeWrapper_) {
-    Blockly.unbindEvent_(this.onMouseDownResizeWrapper_);
+    Blockly.eventHandling.unbind(this.onMouseDownResizeWrapper_);
   }
   Blockly.Bubble.unbindDragEvents_();
   Blockly.utils.dom.removeNode(this.bubbleGroup_);
